requirements.txt
pip
openai
streamlit>=1.16.0
python-dotenv
streamlit-chat
